Fuzz make_bubble
